fix: get the worker to worker
